refactor(bot): move track utils out of utils into services
utils/music/trackUtils/ was never a utils module. It is a stateful
service tree: TrackUtils holds a TrackProcessor, which owns a
TrackCacheManager wrapping an LRUCache plus a cleanup timer, and the
barrel exports a module-level singleton, `export const trackUtils =
new TrackUtils()`. Same pattern as every finding in #1968.

Moved wholesale to services/musicManagement/trackUtils/, the
destination that series established. Only getTrackInfo escapes the
directory, so the external surface is four files: queueDisplay.ts and
queueStats.ts plus the two specs, whose jest.mock() string literals
name the path and do not move with an import rewrite. Those literals
were the most common miss in the #1968 slices, and the same trap
applies inside the moved specs, which mock '../titleComparison'.

Both directories sit three levels under src/, so relative depth is
unchanged and only the one sibling import needed rewriting:
'../titleComparison' becomes '../../../utils/music/titleComparison'.
That is services depending on utils, the direction that was already
correct; the inversion this issue is about is utils depending on
services. titleComparison is itself on the #1991 list and moves later.

Part of #1991. Four subdirectories remain: search/engineManager,
search/providerHealth, titleComparison/service and
youtubeErrorHandler/analyzer.
