+++ packages/bot/src/services/musicManagement/trackUtils/cacheManager.spec.ts
