+++ packages/bot/src/services/musicManagement/trackUtils/cacheManager.ts
