+++ packages/bot/src/services/musicManagement/trackUtils/index.spec.ts
@@ -19,7 +19,7 @@ jest.mock('@lucky/shared/utils', () => ({
     debugLog: jest.fn(),
 }))
 
-jest.mock('../titleComparison', () => ({
+jest.mock('../../../utils/music/titleComparison', () => ({
     isSimilarTitle: jest.fn(),
 }))
 
@@ -47,7 +47,9 @@ function createMockTrack(
 
 describe('standalone function exports', () => {
     beforeEach(() => {
-        const { isSimilarTitle } = require('../titleComparison')
+        const {
+            isSimilarTitle,
+        } = require('../../../utils/music/titleComparison')
         isSimilarTitle.mockReturnValue(false)
     })
 
@@ -172,7 +174,9 @@ describe('TrackUtils class', () => {
 
     describe('findSimilarTracks', () => {
         test('finds tracks with similar titles', () => {
-            const { isSimilarTitle } = require('../titleComparison')
+            const {
+                isSimilarTitle,
+            } = require('../../../utils/music/titleComparison')
             isSimilarTitle.mockImplementation(
                 (title: string, query: string) => {
                     return title.toLowerCase().includes(query.toLowerCase())
@@ -191,7 +195,9 @@ describe('TrackUtils class', () => {
         })
 
         test('respects limit parameter', () => {
-            const { isSimilarTitle } = require('../titleComparison')
+            const {
+                isSimilarTitle,
+            } = require('../../../utils/music/titleComparison')
             isSimilarTitle.mockReturnValue(true)
 
             const tracks = Array.from({ length: 10 }, (_, i) =>
@@ -204,7 +210,9 @@ describe('TrackUtils class', () => {
         })
 
         test('uses default limit of 5', () => {
-            const { isSimilarTitle } = require('../titleComparison')
+            const {
+                isSimilarTitle,
+            } = require('../../../utils/music/titleComparison')
             isSimilarTitle.mockReturnValue(true)
 
             const tracks = Array.from({ length: 10 }, (_, i) =>
@@ -217,7 +225,9 @@ describe('TrackUtils class', () => {
         })
 
         test('returns empty array if no matches', () => {
-            const { isSimilarTitle } = require('../titleComparison')
+            const {
+                isSimilarTitle,
+            } = require('../../../utils/music/titleComparison')
             isSimilarTitle.mockReturnValue(false)
 
             const tracks = [
