+++ packages/bot/src/services/musicManagement/trackUtils/types.ts
